formula_cellar_checks: port to generic OS.
