+++ Library/Homebrew/extend/os/formula_cellar_checks.rb
@@ -1,0 +1,5 @@
+require "formula_cellar_checks"
+
+if OS.mac?
+  require "extend/os/mac/formula_cellar_checks"
+end
